thermal: msm_lmh_dcvs: Never initialize LMH TZ more than once
The LMH DCVS initialization consists in enabling the LMH thermal,
outer loop, reliability and BCL algorithms on a per-cluster
basis, but the command that kickstarts it is "PROFILE_CHANGE"
with parameter '1' sent to the TZ LMH Service.

Sending it more than one time is rendundant, but the critical
issue is that - if we send it more than once - the CPUs will
deadlock!

Change-Id: Id27a65e2890dd4a2128cef4d97312d3b90b018ce
Signed-off-by: savie <ntop_one@yahoo.co.id>