msm: kgsl: Reduce latency while processing ioctls
The GPU's ioctls sleep a lot due to memory allocation, during which time
the CPU can enter a deep idle state and take a while to finish executing
the ioctl afterwards. We can alleviate this by optimistically assuming
that the task running the ioctl won't migrate CPUs while it runs and
restricting the current CPU to shallow idle states.

Change-Id: I15c370657d0b88daba6c97cbabe16feb93006c4f
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: savie <ntop_one@yahoo.co.id>